Added page caching for a significant increase in performance. (Cache uses each page's underlying last file modification time [ctime] as an identifier).
Removed the sha1 password caching implementation and replaced it with an encrypted session cookie. Passwords are stored using a bcrypt hash in the config.json file.

The config.json file is validated upon startup. Plaintext passwords are hashed with bcrypt. Usernames are converted to lowercase. If the SECURITY_KEY is not a hex-encoded 24 byte value, it is replaced with a new random value.